pyproject.toml
[tool.poetry]
name = "transport"
version = "0.1.0"
description = ""
authors = ["MBW <matthew@mh-white.com>"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
loguru = "^0.7.2"


[tool.poetry.group.dev.dependencies]
grpcio-tools = "^1.64.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
